+++ src/main/java/me/primaryuan/carpet/util/SleepUtil.java
@@ -1,8 +1,0 @@
-package me.primaryuan.carpet.util;
-
-import me.primaryuan.carpet.CarpetPrimaryuanSettings;
-
-public class SleepUtil {
-    public static void init() {
-    }
-}